oop.library: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@38156 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/oop/support.c
@@ -156,7 +156,7 @@ BOOL init_methodbase(STRPTR interface_id, ULONG methodbase, ULONG *methodbase_pt
 		*methodbase_ptr = methodbase;
 		
 		/* Leave attrbase field unitialized */
-		idb->attrbase = -1UL;
+		idb->attrbase = (ULONG)-1;
 		
 		/* Insert bucket into hash table */
 		InsertBucket(iidtable, (struct Bucket *)idb, (struct IntOOPBase *)OOPBase);
